requirements.txt
APScheduler==3.7.0
cassandra-driver==3.25.0
certifi==2021.5.30
click==7.1.2
colorama==0.4.4
colorhash==1.0.3
configparser==5.0.2
Flask==1.1.4
Flask-Cors==3.0.10
Flask-MonitoringDashboard
geomet==0.2.1.post1
greenlet==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==1.0.1
lightgbm==3.2.1
MarkupSafe==2.0.1
numpy
pandas
psutil==5.8.0
python-dateutil==2.8.2
pytz==2021.1
pytest==6.2.2
scikit-learn
scipy
six==1.16.0
sklearn
SQLAlchemy==1.4.23
threadpoolctl==2.2.0
tzlocal==2.1
Werkzeug==1.0.1
wincertstore==0.2
xgboost